pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "AI Travel Agent with RAG pipeline using LangChain + LangGraph for grounded recommendations"
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"
dependencies = [
    # API & Web
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "httpx>=0.27",
    "tenacity>=8.2",
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    # Vector store & embeddings
    "chromadb>=0.5.0",
    "tiktoken>=0.7.0",
    # LLM provider
    "openai>=1.40",
    # CLI & UI
    "typer>=0.12",
    "rich>=13.0",
    "streamlit>=1.38",
    "python-dotenv>=1.0",
    # Text processing
    "langchain-text-splitters>=0.3.0",
    "langchain-chroma>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=4.1",
    "httpx>=0.27",  # For FastAPI test client
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
asyncio_mode = "auto"
